Move Layout to commonMain.
